create_API_get_category_and_supplier_fix_book_route

+++ src/migrations/migrate-book.js
@@ -26,7 +26,7 @@ module.exports = {
         type: Sequelize.DOUBLE
       },
       discount: {
-        type: Sequelize.DOUBLE
+        type: Sequelize.INTEGER
       },
       stock: {
         allowNull: false,
